+++ CHANGELOG.md
@@ -3,7 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
-## [0.2.1](https://github.com/chrizzFTD/naming/releases/tag/Release-0.2.0) - 2019-05-04
+## [0.3.0](https://github.com/chrizzFTD/naming/releases/tag/Release-0.3.0) - 2019-05-06
+### Changed
+- `BaseName` renamed to just `Name` on `naming` module.
+- `Name`, `Pipe` and `File` no longer have a `base` field from their config.
+
+## [0.2.1](https://github.com/chrizzFTD/naming/releases/tag/Release-0.2.1) - 2019-05-04
 ### Changed
 - `BaseName` object is now exposed on main `naming` namespace.